requirements.txt
amqp==5.3.1
ansicon==1.89.0
arrow==1.3.0
asgiref==3.9.1
billiard==4.2.1
black==25.1.0
blessed==1.21.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.4
django-celery-beat==2.8.1
django-picklefield==3.3
django-timezone-field==7.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
jinxed==1.3.0
kombu==5.5.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
PyJWT==2.10.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==4.5.5
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20250708
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
